move logic to write file to disk into AemArticleFileManager
