+++ ui/article-aem-renderer/src/test/java/org/cru/godtools/article/aem/service/AemArticleManagerTest.kt
@@ -3,10 +3,9 @@ package org.cru.godtools.article.aem.service
 import android.net.Uri
 import androidx.room.InvalidationTracker
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import java.security.MessageDigest
+import java.io.File
 import java.util.Date
 import kotlin.io.path.ExperimentalPathApi
-import kotlin.io.path.createTempDirectory
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.consumeAsFlow
@@ -15,11 +14,9 @@ import kotlinx.coroutines.test.TestCoroutineScope
 import okhttp3.MediaType
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.ResponseBody.Companion.toResponseBody
-import org.ccci.gto.android.common.kotlin.coroutines.ReadWriteMutex
 import org.cru.godtools.article.aem.api.AemApi
 import org.cru.godtools.article.aem.db.ArticleRoomDatabase
 import org.cru.godtools.article.aem.model.Resource
-import org.cru.godtools.article.aem.util.AemFileSystem
 import org.cru.godtools.base.tool.service.ManifestManager
 import org.cru.godtools.model.Translation
 import org.cru.godtools.tool.model.Manifest
@@ -27,12 +24,8 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.allOf
 import org.hamcrest.Matchers.greaterThanOrEqualTo
 import org.hamcrest.Matchers.lessThanOrEqualTo
-import org.hamcrest.Matchers.startsWith
 import org.junit.After
-import org.junit.Assert.assertArrayEquals
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNotEquals
-import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
@@ -40,7 +33,6 @@ import org.junit.runner.RunWith
 import org.keynote.godtools.android.db.GodToolsDao
 import org.mockito.Mockito.CALLS_REAL_METHODS
 import org.mockito.Mockito.RETURNS_DEEP_STUBS
-import org.mockito.Mockito.mockStatic
 import org.mockito.Mockito.verifyNoInteractions
 import org.mockito.kotlin.UseConstructor
 import org.mockito.kotlin.any
@@ -50,7 +42,6 @@ import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
-import org.mockito.kotlin.spy
 import org.mockito.kotlin.stub
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyBlocking
@@ -61,12 +52,9 @@ import retrofit2.Response
 @RunWith(AndroidJUnit4::class)
 @OptIn(ExperimentalCoroutinesApi::class, ExperimentalPathApi::class)
 class AemArticleManagerTest {
-    private val testDir = createTempDirectory().toFile()
-
     private lateinit var aemDb: ArticleRoomDatabase
     private lateinit var api: AemApi
     private lateinit var dao: GodToolsDao
-    private lateinit var fs: AemFileSystem
     private lateinit var fileManager: AemArticleManager.FileManager
     private lateinit var manifestManager: ManifestManager
     private lateinit var testScope: TestCoroutineScope
@@ -85,17 +73,13 @@ class AemArticleManagerTest {
         dao = mock {
             on { getAsFlow(QUERY_ARTICLE_TRANSLATIONS) } doReturn articleTranslationsChannel.consumeAsFlow()
         }
-        fs = spy(AemFileSystem(mock { on { filesDir } doReturn testDir }))
-        fileManager = mock {
-            // TODO: this should be temporary until we finish refactoring
-            on { filesystemMutex } doReturn ReadWriteMutex()
-        }
+        fileManager = mock()
         manifestManager = mock()
         testScope = TestCoroutineScope().apply { pauseDispatcher() }
 
         articleManager = mock(
             defaultAnswer = CALLS_REAL_METHODS,
-            useConstructor = UseConstructor.withArguments(aemDb, api, dao, fs, fileManager, manifestManager, testScope)
+            useConstructor = UseConstructor.withArguments(aemDb, api, dao, fileManager, manifestManager, testScope)
         )
     }
 
@@ -104,7 +88,6 @@ class AemArticleManagerTest {
         articleTranslationsChannel.close()
         runBlocking { articleManager.shutdown() }
         testScope.cleanupTestCoroutines()
-        testDir.deleteRecursively()
     }
 
     // region Translations
@@ -192,8 +175,10 @@ class AemArticleManagerTest {
         val uri = mock<Uri>()
         val mediaType = "image/jpg".toMediaType()
         val resource = mock<Resource> { on { needsDownload() } doReturn true }
+        val file = mock<File> { on { name } doReturn "file.ext" }
         whenever(resourceDao.find(uri)).thenReturn(resource)
         wheneverDownloadingResource(uri).thenReturn(Response.success(data.toByteArray().toResponseBody(mediaType)))
+        whenever(fileManager.writeFileToDisk(any())) doReturn file
 
         val startTime = System.currentTimeMillis()
         articleManager.downloadResource(uri, false)
@@ -207,8 +192,7 @@ class AemArticleManagerTest {
         verify(resourceDao).updateLocalFile(eq(uri), type.capture(), fileName.capture(), date.capture())
         assertEquals(mediaType, type.firstValue)
         assertThat(date.firstValue.time, allOf(greaterThanOrEqualTo(startTime), lessThanOrEqualTo(endTime)))
-        val file = fs.file(fileName.firstValue)
-        assertArrayEquals(data.toByteArray(), file.readBytes())
+        assertEquals("file.ext", fileName.firstValue)
     }
 
     @Test
@@ -232,43 +216,6 @@ class AemArticleManagerTest {
     }
 
     private fun wheneverDownloadingResource(uri: Uri = any()) = runBlocking { whenever(api.downloadResource(uri)) }
-
-    // region InputStream.writeToDisk()
-    @Test
-    fun testWriteToDisk() = runBlocking {
-        val data = "testWriteToDisk()"
-
-        val file = with(articleManager) { data.byteInputStream().use { it.writeToDisk()!! } }
-        assertNotNull(file)
-        assertArrayEquals(data.toByteArray(), file.readBytes())
-    }
-
-    @Test
-    fun testWriteToDiskDedup() = runBlocking {
-        val data = "testWriteToDiskDedup()"
-
-        val file1 = with(articleManager) { data.byteInputStream().use { it.writeToDisk()!! } }
-        val file2 = with(articleManager) { data.byteInputStream().use { it.writeToDisk()!! } }
-
-        assertEquals(file1, file2)
-    }
-
-    @Test
-    fun testWriteToDiskNoDedupWithoutDigest() = runBlocking {
-        mockStatic(MessageDigest::class.java).use {
-            it.`when`<MessageDigest?> { MessageDigest.getInstance("SHA-1") } doReturn null
-            val data = "testWriteToDiskNoDedupWithoutDigest()"
-
-            val file1 = with(articleManager) { data.byteInputStream().use { it.writeToDisk()!! } }
-            val file2 = with(articleManager) { data.byteInputStream().use { it.writeToDisk()!! } }
-
-            assertNotEquals(file1, file2)
-            assertNotEquals(file1.name, file2.name)
-            assertThat(file1.name, startsWith("aem-"))
-            assertThat(file2.name, startsWith("aem-"))
-        }
-    }
-    // endregion InputStream.writeToDisk()
     // endregion Download Resource
 
     // region cleanupActor
